[OSSM-6799] Fix TestSMCPMultiple after changes in validatingwebhookconfiguration

+++ pkg/tests/non-dependant/bug_multiple_smcp_test.go
@@ -2,6 +2,7 @@ package non_dependant
 
 import (
 	_ "embed"
+	"fmt"
 	"testing"
 
 	"github.com/maistra/maistra-test-tool/pkg/tests/ossm"
@@ -10,7 +11,9 @@ import (
 	"github.com/maistra/maistra-test-tool/pkg/util/oc"
 	"github.com/maistra/maistra-test-tool/pkg/util/pod"
 	"github.com/maistra/maistra-test-tool/pkg/util/retry"
+	"github.com/maistra/maistra-test-tool/pkg/util/shell"
 	. "github.com/maistra/maistra-test-tool/pkg/util/test"
+	"github.com/maistra/maistra-test-tool/pkg/util/version"
 )
 
 func TestSMCPMultiple(t *testing.T) {
@@ -18,49 +21,70 @@ func TestSMCPMultiple(t *testing.T) {
 		t.Log("This test verifies whether the operator only reconciles one SMCP when two exist in a namespace")
 		t.Log("See https://issues.redhat.com/browse/OSSM-2419")
 
-		smcp1 := ossm.DefaultSMCP().WithName("smcp1")
-		smcp2 := ossm.DefaultSMCP().WithName("smcp2")
-
 		t.Cleanup(func() {
 			t.LogStepf("Delete namespace %s", meshNamespace)
 			oc.RecreateNamespace(t, meshNamespace)
 
-			t.LogStep("Delete operator to recreate the ValidationWebhookConfiguration")
-			oc.DeletePod(t, pod.MatchingSelector("name=istio-operator", env.GetOperatorNamespace()))
-
-			t.LogStep("Wait for operator pod to be ready")
-			oc.WaitPodReady(t, pod.MatchingSelector("name=istio-operator", env.GetOperatorNamespace()))
-
-			t.LogStep("Check whether ValidatingWebhookConfiguration exists")
-			retry.UntilSuccess(t, func(t TestHelper) {
-				oc.Get(t, "", "validatingwebhookconfiguration", "openshift-operators.servicemesh-resources.maistra.io")
-				t.LogSuccess("ValidatingWebhookConfiguration was recreated by the operator")
-			})
+			if env.GetOperatorVersion().LessThan(version.OPERATOR_2_6_0) {
+				t.LogStep("Delete operator to recreate the ValidationWebhookConfiguration")
+				oc.DeletePod(t, pod.MatchingSelector("name=istio-operator", env.GetOperatorNamespace()))
+				t.LogStep("Wait for operator pod to be ready")
+				oc.WaitPodReady(t, pod.MatchingSelector("name=istio-operator", env.GetOperatorNamespace()))
+				t.LogStep("Check whether ValidatingWebhookConfiguration exists")
+				retry.UntilSuccess(t, func(t TestHelper) {
+					oc.Get(t, "", "validatingwebhookconfiguration", "openshift-operators.servicemesh-resources.maistra.io")
+					t.LogSuccess("ValidatingWebhookConfiguration was recreated by the operator")
+				})
+			}
 		})
 
 		t.LogStepf("Delete and recreate namespace %s", meshNamespace)
 		oc.RecreateNamespace(t, meshNamespace)
 
-		t.LogStep("Delete the operator's ValidationWebhookConfiguration")
-		oc.DeleteResource(t, "", "validatingwebhookconfiguration", "openshift-operators.servicemesh-resources.maistra.io")
-
 		t.LogStep("Create the first SMCP")
+		smcp1 := ossm.DefaultSMCP().WithName("smcp1")
 		ossm.InstallSMCPCustom(t, meshNamespace, smcp1)
 
 		t.LogStep("Check whether the first SMCP gets reconciled and becomes ready")
 		oc.WaitSMCPReady(t, meshNamespace, smcp1.Name)
 		t.LogSuccess("First SMCP is ready")
 
-		t.LogStep("Create the second SMCP")
-		ossm.InstallSMCPCustom(t, meshNamespace, smcp2)
+		if env.GetOperatorVersion().GreaterThanOrEqual(version.OPERATOR_2_6_0) {
+			t.LogStep("Check that the validationwebhook prevent of creating the second SMCP")
+			shell.Execute(t,
+				fmt.Sprintf("echo \"%s\" | oc create -f - -n %s || true", simpleSmcp(), meshNamespace),
+				assert.OutputContains("admission webhook \"smcp.validation.maistra.io\" denied the request: only one service mesh may be installed per project/namespace",
+					"Validationwebhook prevents of creating the second SMCP",
+					"Expect that creation of second smcp fails on validationwebhook"))
+		} else {
+			t.LogStep("Delete the operator's ValidationWebhookConfiguration to be able to install the second SMCP")
+			oc.DeleteResource(t, "", "validatingwebhookconfiguration", "openshift-operators.servicemesh-resources.maistra.io")
+			t.LogStep("Create the second SMCP")
+			smcp2 := ossm.DefaultSMCP().WithName("smcp2")
+			ossm.InstallSMCPCustom(t, meshNamespace, smcp2)
+			ossm.InstallSMCPCustom(t, meshNamespace, smcp2)
 
-		t.LogStep("Check whether the second SMCP shows ErrMultipleSMCPs")
-		retry.UntilSuccess(t, func(t TestHelper) {
-			oc.Get(t, meshNamespace,
-				"smcp", smcp2.Name,
-				assert.OutputContains("ErrMultipleSMCPs",
-					"The second SMCP status is ErrMultipleSMCPs",
-					"The second SMCP status is not ErrMultipleSMCPs"))
-		})
+			ossm.InstallSMCPCustom(t, meshNamespace, smcp2)
+
+			t.LogStep("Check whether the second SMCP shows ErrMultipleSMCPs")
+			retry.UntilSuccess(t, func(t TestHelper) {
+				oc.Get(t, meshNamespace,
+					"smcp", smcp2.Name,
+					assert.OutputContains("ErrMultipleSMCPs",
+						"The second SMCP status is ErrMultipleSMCPs",
+						"The second SMCP status is not ErrMultipleSMCPs"))
+			})
+		}
 	})
+}
+
+func simpleSmcp() string {
+	return fmt.Sprintf(`
+kind: ServiceMeshControlPlane
+apiVersion: maistra.io/v2
+metadata:
+  name: basic
+spec: 
+  version: %s
+`, env.GetSMCPVersion())
 }
